requirements.txt
python-decouple==3.3
wget==3.2
